requirements.txt
torch==2.5.1
torchvision==0.20.1
transformers==4.48.2
flask
Pillow==11.1.0
matplotlib==3.10.0